Refactor api, relocate routes and clean up the code
