+++ lib/utils.js
@@ -36,18 +36,14 @@ function validate_name(name) {
 	name = name.toLowerCase();
 
 	// all URL-safe characters and "@" for issue #75
-	if (!name.match(/^[-a-zA-Z0-9_.!~*'()@]+$/)
+	return !(!name.match(/^[-a-zA-Z0-9_.!~*'()@]+$/)
 	 || name.charAt(0) === '.' // ".bin", etc.
 	 || name.charAt(0) === '-' // "-" is reserved by couchdb
 	 || name === 'node_modules'
 	 || name === '__proto__'
 	 || name === 'package.json'
 	 || name === 'favicon.ico'
-	) {
-		return false;
-	} else {
-		return true;
-	}
+	);
 }
 
 /**
