+++ lib/web/api/endpoint/ping.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = function(route) {
+  route.get('/-/ping', function(req, res, next) {
+    next({});
+  });
+};
