+++ lib/web/index.js
@@ -1,55 +1,57 @@
 'use strict';
 
-let async = require('async');
-let bodyParser = require('body-parser');
-let Cookies = require('cookies');
-let express = require('express');
-let fs = require('fs');
-let Handlebars = require('handlebars');
-let renderReadme = require('render-readme');
-let Search = require('../search');
-let Middleware = require('./middleware');
-let Utils = require('../utils');
-let match = Middleware.match;
-let validate_name = Middleware.validate_name;
-let validate_pkg = Middleware.validate_package;
+const async = require('async');
+const bodyParser = require('body-parser');
+const Cookies = require('cookies');
+const escape = require('js-string-escape');
+const express = require('express');
+const fs = require('fs');
+const Handlebars = require('handlebars');
+const marked = require('marked');
+const Search = require('../search');
+const Middleware = require('./middleware');
+const Utils = require('../utils');
+const match = Middleware.match;
+const validateName = Middleware.validate_name;
+const validatePkg = Middleware.validate_package;
+const securityIframe = Middleware.securityIframe;
 
 module.exports = function(config, auth, storage) {
+  Search.configureStorage(storage);
   /* eslint new-cap:off */
-  let app = express.Router();
-  let can = Middleware.allow(auth);
+  const app = express.Router();
+  /* eslint new-cap:off */
+  const can = Middleware.allow(auth);
+  let template;
 
   // validate all of these params as a package name
   // this might be too harsh, so ask if it causes trouble
-  app.param('package', validate_pkg);
-  app.param('filename', validate_name);
-  app.param('version', validate_name);
+  app.param('package', validatePkg);
+  app.param('filename', validateName);
+  app.param('version', validateName);
   app.param('anything', match(/.*/));
 
   app.use(Cookies.express());
   app.use(bodyParser.urlencoded({extended: false}));
   app.use(auth.cookie_middleware());
-  app.use(function(req, res, next) {
-    // disable loading in frames (clickjacking, etc.)
-    res.header('X-Frame-Options', 'deny');
-    next();
-  });
-
-  Search.configureStorage(storage);
+  app.use(securityIframe);
 
   Handlebars.registerPartial('entry', fs.readFileSync(require.resolve('./ui/entry.hbs'), 'utf8'));
-  let template;
+
   if (config.web && config.web.template) {
     template = Handlebars.compile(fs.readFileSync(config.web.template, 'utf8'));
   } else {
     template = Handlebars.compile(fs.readFileSync(require.resolve('./ui/index.hbs'), 'utf8'));
   }
+
   app.get('/', function(req, res, next) {
     let base = Utils.combineBaseUrl(Utils.getWebProtocol(req), req.get('host'), config.url_prefix);
     res.setHeader('Content-Type', 'text/html');
 
     storage.get_local(function(err, packages) {
-      if (err) throw err; // that function shouldn't produce any
+      if (err) {
+        throw err;
+      } // that function shouldn't produce any
       async.filterSeries(packages, function(pkg, cb) {
         auth.allow_access(pkg.name, req.remote_user, function(err, allowed) {
           setImmediate(function() {
@@ -69,13 +71,15 @@ module.exports = function(config, auth, storage) {
             return 1;
           }
         });
-
-        next(template({
-          name: config.web && config.web.title ? config.web.title : 'Verdaccio',
-          tagline: config.web && config.web.tagline ? config.web.tagline : '',
+        let json = {
           packages: packages,
+          tagline: config.web && config.web.tagline ? config.web.tagline : '',
           baseUrl: base,
           username: req.remote_user.name,
+        };
+        next(template({
+          name: config.web && config.web.title ? config.web.title : 'Verdaccio',
+          data: escape(JSON.stringify(json)),
         }));
       });
     });
@@ -85,7 +89,9 @@ module.exports = function(config, auth, storage) {
   app.get('/-/static/:filename', function(req, res, next) {
     let file = __dirname + '/static/' + req.params.filename;
     res.sendFile(file, function(err) {
-      if (!err) return;
+      if (!err) {
+        return;
+      }
       if (err.status === 404) {
         next();
       } else {
@@ -101,10 +107,9 @@ module.exports = function(config, auth, storage) {
   });
 
   app.post('/-/login', function(req, res, next) {
-    auth.authenticate(req.body.user, req.body.pass, function(err, user) {
+    auth.authenticate(req.body.user, req.body.pass, (err, user) => {
       if (!err) {
         req.remote_user = user;
-        // res.cookies.set('token', auth.issue_token(req.remote_user))
 
         let str = req.body.user + ':' + req.body.pass;
         res.cookies.set('token', auth.aes_encrypt(str).toString('base64'));
@@ -127,10 +132,12 @@ module.exports = function(config, auth, storage) {
     const packages = [];
 
     const getData = function(i) {
-      storage.get_package(results[i].ref, function(err, entry) {
+      storage.get_package(results[i].ref, (err, entry) => {
         if (!err && entry) {
           auth.allow_access(entry.name, req.remote_user, function(err, allowed) { // TODO: This may cause performance issue?
-            if (err || !allowed) return;
+            if (err || !allowed) {
+              return;
+            }
 
             packages.push(entry.versions[entry['dist-tags'].latest]);
           });
@@ -153,10 +160,15 @@ module.exports = function(config, auth, storage) {
 
   app.get('/-/readme(/@:scope?)?/:package/:version?', can('access'), function(req, res, next) {
     let packageName = req.params.package;
-    if (req.params.scope) packageName = '@'+ req.params.scope + '/' + packageName;
+    if (req.params.scope) {
+      packageName = `@${req.params.scope}/${packageName}`;
+    }
     storage.get_package(packageName, {req: req}, function(err, info) {
-      if (err) return next(err);
-      next( renderReadme(info.readme || 'ERROR: No README data found!') );
+      if (err) {
+        return next(err);
+      }
+      res.set('Content-Type', 'text/plain');
+      next( marked(info.readme || 'ERROR: No README data found!') );
     });
   });
   return app;
