+++ lib/web/middleware.js
@@ -3,18 +3,26 @@
 'use strict';
 
 const crypto = require('crypto');
-const Error = require('http-errors');
+const _ = require('lodash');
+const createError = require('http-errors');
 const utils = require('../utils');
 const Logger = require('../logger');
 
+
 module.exports.match = function match(regexp) {
-  return function(req, res, next, value, name) {
+  return function(req, res, next, value) {
     if (regexp.exec(value)) {
       next();
     } else {
       next('route');
     }
   };
+};
+
+module.exports.securityIframe = function securityIframe(req, res, next) {
+  // disable loading in frames (clickjacking, etc.)
+  res.header('X-Frame-Options', 'deny');
+  next();
 };
 
 module.exports.validate_name = function validate_name(req, res, next, value, name) {
@@ -24,7 +32,7 @@ module.exports.validate_name = function validate_name(req, res, next, value, nam
   } else if (utils.validate_name(value)) {
     next();
   } else {
-    next( Error[403]('invalid ' + name) );
+    next( createError[403]('invalid ' + name) );
   }
 };
 
@@ -35,24 +43,32 @@ module.exports.validate_package = function validate_package(req, res, next, valu
   } else if (utils.validate_package(value)) {
     next();
   } else {
-    next( Error[403]('invalid ' + name) );
+    next( createError[403]('invalid ' + name) );
   }
 };
 
 module.exports.media = function media(expect) {
   return function(req, res, next) {
     if (req.headers['content-type'] !== expect) {
-      next( Error[415]('wrong content-type, expect: ' + expect
-                     + ', got: '+req.headers['content-type']) );
+      next( createError[415]('wrong content-type, expect: ' + expect
+        + ', got: '+req.headers['content-type']) );
     } else {
       next();
     }
   };
 };
 
+module.exports.encodeScopePackage = function(req, res, next) {
+  if (req.url.indexOf('@') !== -1) {
+    // e.g.: /@org/pkg/1.2.3 -> /@org%2Fpkg/1.2.3, /@org%2Fpkg/1.2.3 -> /@org%2Fpkg/1.2.3
+    req.url = req.url.replace(/^(\/@[^\/%]+)\/(?!$)/, '$1%2F');
+  }
+  next();
+};
+
 module.exports.expect_json = function expect_json(req, res, next) {
   if (!utils.is_object(req.body)) {
-    return next( Error[400]('can\'t parse incoming json') );
+    return next( createError[400]('can\'t parse incoming json') );
   }
   next();
 };
@@ -65,7 +81,7 @@ module.exports.anti_loop = function(config) {
       for (let i=0; i<arr.length; i++) {
         let m = arr[i].match(/\s*(\S+)\s+(\S+)/);
         if (m && m[2] === config.server_id) {
-          return next( Error[508]('loop detected') );
+          return next( createError[508]('loop detected') );
         }
       }
     }
@@ -84,11 +100,12 @@ function md5sum(data) {
   return crypto.createHash('md5').update(data).digest('hex');
 }
 
+
 module.exports.allow = function(auth) {
   return function(action) {
     return function(req, res, next) {
       req.pause();
-      auth['allow_'+action](req.params.package, req.remote_user, function(error, allowed) {
+      auth['allow_' + action](req.params.package, req.remote_user, function(error, allowed) {
         req.resume();
         if (error) {
           next(error);
@@ -97,7 +114,7 @@ module.exports.allow = function(auth) {
         } else {
           // last plugin (that's our built-in one) returns either
           // cb(err) or cb(null, true), so this should never happen
-          throw Error('bug in the auth plugin system');
+          throw createError('bug in the auth plugin system');
         }
       });
     };
@@ -111,12 +128,12 @@ module.exports.final = function(body, req, res, next) {
   }
 
   try {
-    if (typeof(body) === 'string' || typeof(body) === 'object') {
+    if (_.isString(body) || _.isObject(body)) {
       if (!res.getHeader('Content-type')) {
         res.header('Content-type', 'application/json');
       }
 
-      if (typeof(body) === 'object' && body != null) {
+      if (typeof(body) === 'object' && _.isNil(body) === false) {
         if (typeof(body.error) === 'string') {
           res._verdaccio_error = body.error;
         }
@@ -135,7 +152,9 @@ module.exports.final = function(body, req, res, next) {
     // as an error handler, we can't report error properly,
     // and should just close socket
     if (err.message.match(/set headers after they are sent/)) {
-      if (res.socket != null) res.socket.destroy();
+      if (_.isNil(res.socket) === false) {
+        res.socket.destroy();
+      }
       return;
     } else {
       throw err;
@@ -150,17 +169,26 @@ module.exports.log = function(req, res, next) {
   req.log = Logger.logger.child({sub: 'in'});
 
   let _auth = req.headers.authorization;
-  if (_auth != null) req.headers.authorization = '<Classified>';
+  if (_.isNil(_auth) === false) {
+    req.headers.authorization = '<Classified>';
+  }
   let _cookie = req.headers.cookie;
-  if (_cookie != null) req.headers.cookie = '<Classified>';
+  if (_.isNil(_cookie) === false) {
+    req.headers.cookie = '<Classified>';
+  }
 
   req.url = req.originalUrl;
   req.log.info( {req: req, ip: req.ip}
-              , '@{ip} requested \'@{req.method} @{req.url}\'' );
+    , '@{ip} requested \'@{req.method} @{req.url}\'' );
   req.originalUrl = req.url;
 
-  if (_auth != null) req.headers.authorization = _auth;
-  if (_cookie != null) req.headers.cookie = _cookie;
+  if (_.isNil(_auth) === false) {
+    req.headers.authorization = _auth;
+  }
+
+  if (_.isNil(_cookie) === false) {
+    req.headers.cookie = _cookie;
+  }
 
   let bytesin = 0;
   req.on('data', function(chunk) {
@@ -187,7 +215,10 @@ module.exports.log = function(req, res, next) {
 
     req.url = req.originalUrl;
     req.log.warn({
-      request: {method: req.method, url: req.url},
+      request: {
+        method: req.method,
+        url: req.url,
+      },
       level: 35, // http
       user: req.remote_user && req.remote_user.name,
       remoteIP,
@@ -205,9 +236,11 @@ module.exports.log = function(req, res, next) {
     log(true);
   });
 
-  let _end = res.end;
+  const _end = res.end;
   res.end = function(buf) {
-    if (buf) bytesout += buf.length;
+    if (buf) {
+      bytesout += buf.length;
+    }
     _end.apply(res, arguments);
     log();
   };
