+++ lib/web/utils/const.js
@@ -1,0 +1,5 @@
+'use strict';
+
+const CONTENT_JSON = 'application/json';
+
+module.exports.CONTENT_JSON = CONTENT_JSON;
